Change gfloat to gdouble in image display
There is some inconsistent use of gfloat vs gdouble in files.
Switch to using gdouble in the image display to address this.
Bug #20122

+++ src/frontend/entangle-image-display.h
@@ -69,16 +69,16 @@ gboolean entangle_image_display_get_autoscale(EntangleImageDisplay *display);
 
 
 void entangle_image_display_set_scale(EntangleImageDisplay *display,
-                                      gfloat scale);
-gfloat entangle_image_display_get_scale(EntangleImageDisplay *display);
+                                      gdouble scale);
+gdouble entangle_image_display_get_scale(EntangleImageDisplay *display);
 
 void entangle_image_display_set_aspect_ratio(EntangleImageDisplay *display,
-                                             gfloat aspect);
-gfloat entangle_image_display_get_aspect_ratio(EntangleImageDisplay *display);
+                                             gdouble aspect);
+gdouble entangle_image_display_get_aspect_ratio(EntangleImageDisplay *display);
 
 void entangle_image_display_set_mask_opacity(EntangleImageDisplay *display,
-                                             gfloat opacity);
-gfloat entangle_image_display_get_mask_opacity(EntangleImageDisplay *display);
+                                             gdouble opacity);
+gdouble entangle_image_display_get_mask_opacity(EntangleImageDisplay *display);
 
 void entangle_image_display_set_mask_enabled(EntangleImageDisplay *display,
                                              gboolean enabled);
